Extract value and validator modules
