+++ esp-config/src/lib.rs
@@ -8,7 +8,7 @@
 #[cfg(feature = "build")]
 mod generate;
 #[cfg(feature = "build")]
-pub use generate::{generate_config, Error, Validator, Value};
+pub use generate::{generate_config, validator::Validator, value::Value, Error};
 
 /// Parse the value of an environment variable as a [bool] at compile time.
 #[macro_export]
